rename file for meal migration
